+++ src/app/preview-registry.ts
@@ -10,12 +10,14 @@ import { buttonsPreviewModule } from '../plugins/basic-buttons';
 import { cardPreviewModule } from '../plugins/basic-card';
 import { inputsPreviewModule } from '../plugins/basic-inputs';
 import { modalPreviewModule } from '../plugins/basic-modal';
+import { radiusPreviewModule } from '../plugins/basic-radius';
 import { tablePreviewModule } from '../plugins/basic-table';
 
 export const previewModules = [
   buttonsPreviewModule,
   inputsPreviewModule,
   cardPreviewModule,
+  radiusPreviewModule,
   alertPreviewModule,
   tablePreviewModule,
   modalPreviewModule,